requirements.txt
protobuf
grpcio
grpcio-tools
torch==2.3.1
torchvision==0.18.1
numpy>=1.26.0
python-dotenv==1.0.0
grpcio-reflection==1.59.3
h5py>=3.10.0
pytest>=7.4.0
pytest-cov>=4.1.0
pika>=1.3.2
psycopg2-binary>=2.9.9 
